requirements.txt
python-telegram-bot==21.6
gspread
google-auth